+++ src/Files.App/Data/EventArguments/PageTypeUpdatedEventArgs.cs
@@ -1,0 +1,9 @@
+﻿namespace Files.App.Data.EventArguments
+{
+	public class PageTypeUpdatedEventArgs
+	{
+		public bool IsTypeCloudDrive { get; set; }
+
+		public bool IsTypeRecycleBin { get; set; }
+	}
+}
