+++ src/Files.App/ViewModels/UserControls/ToolbarViewModel.cs
