+++ src/Files.Backend/Enums/PreviewPaneStates.cs
@@ -1,0 +1,11 @@
+﻿namespace Files.Backend.Enums
+{
+	public enum PreviewPaneStates
+	{
+		NoItemSelected,
+		NoPreviewAvailable,
+		NoPreviewOrDetailsAvailable,
+		PreviewAndDetailsAvailable,
+		LoadingPreview,
+	}
+}
